feat: add possibility for automatic creation of CRUD procedures per table
