+++ sql/index.js
@@ -1,0 +1,10 @@
+module.exports = {
+    auditLog: require('./auditLog'),
+    callParams: require('./callParams'),
+    createDatabase: require('./createDatabase'),
+    createUser: require('./createUser'),
+    disableDatabaseDiagrams: require('./disableDatabaseDiagrams'),
+    enableDatabaseDiagrams: require('./enableDatabaseDiagrams'),
+    loadSchema: require('./loadSchema'),
+    refreshView: require('./refreshView')
+};
